Move websockets to network directory
