+++ app/src/main/java/com/example/traders/profile/portfolio/PortfolioViewModel.kt
@@ -4,7 +4,7 @@ import com.example.traders.BaseViewModel
 import com.example.traders.database.Crypto
 import com.example.traders.profile.cryptoData.CryptoInUsd
 import com.example.traders.profile.cryptoData.CryptoTicker
-import com.example.traders.repository.CryptoRepository
+import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.github.mikephil.charting.data.PieEntry
 import dagger.hilt.android.lifecycle.HiltViewModel
