+++ app/src/test/java/com/example/traders/dialogs/sellDialog/SellDialogViewModelTest.kt
@@ -4,7 +4,7 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import app.cash.turbine.test
 import com.example.traders.database.Crypto
 import com.example.traders.dialogs.DialogValidation
-import com.example.traders.repository.CryptoRepository
+import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.example.traders.TestDispatcherRule
 import com.example.traders.watchlist.cryptoData.FixedCryptoList
