+++ app/src/test/java/com/example/traders/profile/portfolio/PortfolioViewModelTest.kt
@@ -8,7 +8,7 @@ import com.example.traders.TestDispatcherRule
 import com.example.traders.database.Crypto
 import com.example.traders.profile.cryptoData.CryptoInUsd
 import com.example.traders.profile.cryptoData.CryptoTicker
-import com.example.traders.repository.CryptoRepository
+import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.github.mikephil.charting.data.PieData
 import com.github.mikephil.charting.data.PieEntry
